refactor: chat component
